Extracted filteredProducts into useProducts
