+++ src/hooks/useProducts.ts
@@ -6,9 +6,43 @@ import {
   Product,
 } from "../database/database";
 import { sortProducts, SortOrder } from "../utils/sortUtils";
+import { preprocessName, calculateSimilarity } from "../utils/similarityUtils";
+
+
+
+const searchSimilarityThreshold = 0.4;
 
 export default function useProducts(listId: number, sortOrder: SortOrder, searchQuery: string) {
   const [products, setProducts] = useState<Product[]>([]);
+
+  
+    const filteredProducts = products.filter((product) => {
+      const processedProductName = preprocessName(product.name);
+      const processedSearchQuery = preprocessName(searchQuery);
+  
+      // If search query is empty, return all products
+      if (!processedSearchQuery) {
+        return true;
+      }
+  
+      // Calculate the length threshold
+      const nameLength = processedProductName.length;
+      const queryLength = processedSearchQuery.length;
+      const lengthThreshold = Math.ceil(nameLength * 0.5); // 50% of the product name length
+  
+      // Use a simple substring match if the query is less than 50% of the product name length
+      if (queryLength < lengthThreshold) {
+        return processedProductName.includes(processedSearchQuery);
+      }
+  
+      // Calculate similarity if the query meets the length requirement
+      const similarity = calculateSimilarity(
+        processedProductName,
+        processedSearchQuery
+      );
+  
+      return similarity >= searchSimilarityThreshold;
+    });
 
 
   const loadProducts = useCallback(async () => {
@@ -46,6 +80,7 @@ export default function useProducts(listId: number, sortOrder: SortOrder, search
 
   return {
     products,
+    filteredProducts,
     setProducts,
     loadProducts,
     updateQuantity,
